+++ vscode-dotnet-runtime-library/src/Acquisition/DotnetPathFinder.ts
@@ -93,6 +93,23 @@ export class DotnetPathFinder implements IDotnetPathFinder
     }
 
     /**
+     * @remarks This only checks dotnet --list-runtimes or --list-sdks, so it can be more performant for the base case.
+     * This allows skipping which, where, and also does not rely on --info which is slower because it shells to the SDK instead of just being the host.
+     * @returns The path to the dotnet executable, which may be a symlink to the actual executable.
+     */
+    public async findDotnetFastFromListOnly(): Promise<string[] | undefined>
+    {
+        const oldLookup = process.env.DOTNET_MULTILEVEL_LOOKUP;
+        try {
+            process.env.DOTNET_MULTILEVEL_LOOKUP = '0'; // make it so --list-runtimes only finds the runtimes on that path: https://learn.microsoft.com/en-us/dotnet/core/compatibility/deployment/7.0/multilevel-lookup#reason-for-change
+            const finalPath = await this.getTruePath(['dotnet']);
+            return this.returnWithRestoringEnvironment(finalPath, 'DOTNET_MULTILEVEL_LOOKUP', oldLookup);
+        } finally {
+            process.env.DOTNET_MULTILEVEL_LOOKUP = oldLookup; // Ensure the environment variable is always restored
+        }
+    }
+
+    /**
      *
      * @returns A set of the path environment variable(s) for which or where dotnet, which may need to be converted to the actual path if it points to a polymorphic executable.
      * For example, `snap` installs dotnet to snap/bin/dotnet, which you can call --list-runtimes on.
@@ -133,7 +150,8 @@ export class DotnetPathFinder implements IDotnetPathFinder
 
         else
         {
-            this.executor?.execute(CommandExecutor.makeCommand('env', []), { dotnetInstallToolCacheTtlMs: DOTNET_INFORMATION_CACHE_DURATION_MS }, false)
+            this.executor?.execute(CommandExecutor.makeCommand('echo', ['$PATH']), { dotnetInstallToolCacheTtlMs: DOTNET_INFORMATION_CACHE_DURATION_MS }, false)
+
                 .then((result) =>
                 {
                     // Log the default shell state
@@ -143,7 +161,8 @@ export class DotnetPathFinder implements IDotnetPathFinder
 
             if (!(options.shell === '/bin/bash'))
             {
-                this.executor?.execute(CommandExecutor.makeCommand('env', []), { shell: 'bin/bash', dotnetInstallToolCacheTtlMs: SYS_CMD_SEARCH_CACHE_DURATION_MS }, false)
+                this.executor?.execute(CommandExecutor.makeCommand('echo', ['$PATH']), { shell: 'bin/bash', dotnetInstallToolCacheTtlMs: SYS_CMD_SEARCH_CACHE_DURATION_MS }, false)
+
                     .then((result) =>
                     {
                         this.workerContext.eventStream.post(new DotnetFindPathLookupPATH(`Execution Path (Unix Bash): ${result?.stdout}`));