+++ vscode-dotnet-runtime-library/src/EventStream/EventStreamEvents.ts
@@ -1168,6 +1168,16 @@ export class DotnetFindPathNoHostOnRegistry extends DotnetCustomMessageEvent
     public readonly eventName = 'DotnetFindPathNoHostOnRegistry';
 }
 
+export class SudoDirCreationFailed extends DotnetCustomMessageEvent
+{
+    public readonly eventName = 'SudoDirCreationFailed';
+}
+
+export class SudoDirDeletionFailed extends DotnetCustomMessageEvent
+{
+    public readonly eventName = 'SudoDirDeletionFailed';
+}
+
 export class DotnetFindPathOnFileSystem extends DotnetCustomMessageEvent
 {
     public readonly eventName = 'DotnetFindPathOnFileSystem';
@@ -1526,6 +1536,10 @@ export class NetInstallerEndExecutionEvent extends DotnetCustomMessageEvent
     public readonly eventName = 'NetInstallerEndExecutionEvent';
 }
 
+export class FailedToRunSudoCommand extends DotnetCustomMessageEvent
+{
+    public readonly eventName = 'FailedToRunSudoCommand';
+}
 
 export class DotnetInstallLinuxChecks extends DotnetCustomMessageEvent
 {
@@ -1548,51 +1562,13 @@ export abstract class DotnetLockEvent extends DotnetFileEvent
 
     public getProperties()
     {
-        return { Message: this.eventMessage, Time: this.time, Lock: this.lock, File: this.file };
+        return { Message: this.eventMessage, Time: this.time, Lock: this.lock, File: this.file, ...getDisabledTelemetryOnChance(1) };
     }
 }
 
-export class DotnetLockAcquiredEvent extends DotnetLockEvent
+export class GenericDotnetLockEvent extends DotnetLockEvent
 {
-    public readonly eventName = 'DotnetLockAcquiredEvent';
-
-    public getProperties()
-    {
-        return { suppressTelemetry: 'true', ...super.getProperties() };
-    }
-}
-
-export class DotnetLockReleasedEvent extends DotnetLockEvent
-{
-    public readonly eventName = 'DotnetLockReleasedEvent';
-
-    public getProperties()
-    {
-        return { suppressTelemetry: 'true', ...super.getProperties() };
-    }
-}
-
-export class DotnetLockErrorEvent extends DotnetLockEvent
-{
-    public readonly eventName = 'DotnetLockErrorEvent';
-    constructor(public readonly error: Error,
-        public readonly eventMessage: string, public readonly time: string, public readonly lock: string, public readonly file: string) { super(eventMessage, time, lock, file); }
-
-    public getProperties()
-    {
-        return { Error: this.error.toString(), Message: this.eventMessage, Time: this.time, Lock: this.lock, File: this.file };
-    }
-
-}
-
-export class DotnetLockAttemptingAcquireEvent extends DotnetLockEvent
-{
-    public readonly eventName = 'DotnetLockAttemptingAcquireEvent';
-
-    public getProperties()
-    {
-        return { suppressTelemetry: 'true', ...super.getProperties() };
-    }
+    public readonly eventName = 'GenericDotnetLockEvent';
 }
 
 export class DotnetFileWriteRequestEvent extends DotnetFileEvent