+++ vscode-dotnet-runtime-library/src/test/mocks/MockObjects.ts
@@ -474,7 +474,7 @@ export class MockFileUtilities extends IFileUtilities
 
     public async realpath(filePath: string): Promise<string | null>
     {
-        return new FileUtilities().realpath(filePath);
+        return this.filePathsAndReadValues[filePath] || new FileUtilities().realpath(filePath);
     }
 
 }