+++ vscode-dotnet-runtime-library/src/test/unit/LinuxDistroTests.test.ts
@@ -7,18 +7,21 @@ import * as chai from 'chai';
 import * as os from 'os';
 import { DotnetInstallMode } from '../../Acquisition/DotnetInstallMode';
 import { GenericDistroSDKProvider } from '../../Acquisition/GenericDistroSDKProvider';
-import { DistroVersionPair, DotnetDistroSupportStatus, LinuxVersionResolver } from '../../Acquisition/LinuxVersionResolver';
+import { DistroVersionPair, DotnetDistroSupportStatus } from '../../Acquisition/LinuxVersionResolver';
+import * as versionUtils from '../../Acquisition/VersionUtilities';
 import { LocalMemoryCacheSingleton } from '../../LocalMemoryCacheSingleton';
 import { WebRequestWorkerSingleton } from '../../Utils/WebRequestWorkerSingleton';
 import { MockCommandExecutor } from '../mocks/MockObjects';
-import { getMockAcquisitionContext, getMockUtilityContext } from './TestUtility';
+import { UBUNTU_DISTRO_INFO_KEY } from '../../Acquisition/StringConstants';
+import { getLatestLinuxDotnet, getLinuxSupportedDotnetSDKVersion, getMockAcquisitionContext, getMockUtilityContext } from './TestUtility';
+import { getMajor, getMajorMinor } from '../../Acquisition/VersionUtilities';
 const assert = chai.assert;
 const standardTimeoutTime = 100000;
 
-const mockVersion = '7.0.103';
+const mockVersion = '8.0.103';
 const acquisitionContext = getMockAcquisitionContext('sdk', mockVersion);
 const mockExecutor = new MockCommandExecutor(acquisitionContext, getMockUtilityContext());
-const pair: DistroVersionPair = { distro: 'Ubuntu', version: '22.04' };
+const pair: DistroVersionPair = { distro: UBUNTU_DISTRO_INFO_KEY, version: '24.04' };
 const provider: GenericDistroSDKProvider = new GenericDistroSDKProvider(pair, acquisitionContext, getMockUtilityContext(), mockExecutor);
 const shouldRun = os.platform() === 'linux';
 const installType: DotnetInstallMode = 'sdk';
@@ -27,8 +30,8 @@ Command 'dotnet' not found, but can be installed with:
 
             snap install dotnet-sdk # version 7.0.304, or
 
-            apt install dotnet-host # version 6.0.118-06ubuntu1~22.04.1
-            apt install dotnet-host-7.0 # version 7.0.107-6ubuntu1~22.04.1
+            apt install dotnet-host # version 6.0.118-06ubuntu1~24.04.1
+            apt install dotnet-host-7.0 # version 7.0.107-6ubuntu1~24.04.1
             See 'snap info dotnet-sdk' for additional versions.
 `
 
@@ -46,11 +49,12 @@ suite('Linux Distro Logic Unit Tests', function ()
         if (shouldRun)
         {
             const recVersion = await provider.getRecommendedDotnetVersion(installType);
+            const correctVersion = await getLinuxSupportedDotnetSDKVersion(acquisitionContext);
+            const correctXXVersion = `${versionUtils.getMajorMinor(correctVersion, acquisitionContext.eventStream, acquisitionContext)}.1xx`;
             assert.equal(mockExecutor.attemptedCommand,
-                'apt-cache -o DPkg::Lock::Timeout=180 search --names-only ^dotnet-sdk-9.0$', 'Searched for the newest package last with regex'); // this may fail if test not exec'd first
+                `apt-cache -o DPkg::Lock::Timeout=180 search --names-only ^dotnet-sdk-${versionUtils.getMajorMinor(getLatestLinuxDotnet(), acquisitionContext.eventStream, acquisitionContext)}$`, 'Searched for the newest package last with regex'); // this may fail if test not exec'd first
             // the data is cached so --version may not be executed.
-            const distroVersion = await new LinuxVersionResolver(acquisitionContext, getMockUtilityContext()).getRunningDistro();
-            assert.equal(recVersion, Number(distroVersion.version) >= 22.04 ? '9.0.1xx' : '8.0.1xx', 'Resolved the most recent available version : will eventually break if the mock data is not updated');
+            assert.equal(recVersion, correctXXVersion, 'Resolved the most recent available version : will eventually break if the mock data is not updated');
         }
     }).timeout(standardTimeoutTime);
 
@@ -60,7 +64,8 @@ suite('Linux Distro Logic Unit Tests', function ()
         {
             // assert this passes : we don't want the test to be reliant on machine state for whether the package exists or not, so don't check output
             await provider.dotnetPackageExistsOnSystem(mockVersion, installType);
-            assert.equal(mockExecutor.attemptedCommand, 'dpkg -l dotnet-sdk-7.0');
+            const version = await getLinuxSupportedDotnetSDKVersion(acquisitionContext);
+            assert.equal(mockExecutor.attemptedCommand, `dpkg -l dotnet-sdk-${versionUtils.getMajorMinor(version, acquisitionContext.eventStream, acquisitionContext)}`);
         }
     }).timeout(standardTimeoutTime);
 
@@ -144,10 +149,10 @@ Microsoft.NETCore.App 7.0.5 [/usr/lib/dotnet/shared/Microsoft.NETCore.App]`, std
     {
         if (shouldRun)
         {
-            mockExecutor.fakeReturnValue = { stdout: `7.0.105`, stderr: '', status: '0' };
+            mockExecutor.fakeReturnValue = { stdout: mockVersion, stderr: '', status: '0' };
             let currentInfo = await provider.getInstalledGlobalDotnetVersionIfExists();
             mockExecutor.resetReturnValues();
-            assert.equal(currentInfo, '7.0.105');
+            assert.equal(currentInfo, mockVersion);
 
             mockExecutor.fakeReturnValue = { stdout: noDotnetString, stderr: noDotnetString, status: '0' };
             currentInfo = await provider.getInstalledGlobalDotnetVersionIfExists();
@@ -160,13 +165,14 @@ Microsoft.NETCore.App 7.0.5 [/usr/lib/dotnet/shared/Microsoft.NETCore.App]`, std
     {
         if (shouldRun)
         {
-            let supported = await provider.isDotnetVersionSupported('11.0.101', installType);
-            // In the mock data, 8.0 is not supported, so it should be false.
-            assert.equal(supported, false);
-            supported = await provider.isDotnetVersionSupported('7.0.101', installType);
+            const getPlusOneLatest = (Number(getMajor(getLatestLinuxDotnet(), acquisitionContext.eventStream, acquisitionContext)) + 1).toString() + '.0.100';
+            let supported = await provider.isDotnetVersionSupported(getPlusOneLatest, installType);
+            assert.equal(supported, false, '.net x+1 does not exist yet');
+            supported = await provider.isDotnetVersionSupported(await getLinuxSupportedDotnetSDKVersion(acquisitionContext), installType);
             assert.equal(supported, true);
+
             // this feature band isn't supported by most distros yet.
-            supported = await provider.isDotnetVersionSupported('7.0.201', installType);
+            supported = await provider.isDotnetVersionSupported((await getLinuxSupportedDotnetSDKVersion(acquisitionContext)).replace('1', '2'), installType);
             assert.equal(supported, false);
         }
     }).timeout(standardTimeoutTime);
@@ -176,7 +182,7 @@ Microsoft.NETCore.App 7.0.5 [/usr/lib/dotnet/shared/Microsoft.NETCore.App]`, std
         if (shouldRun)
         {
             await provider.installDotnet(mockVersion, installType);
-            assert.equal(mockExecutor.attemptedCommand, 'sudo apt-get -o DPkg::Lock::Timeout=180 install -y dotnet-sdk-7.0');
+            assert.equal(mockExecutor.attemptedCommand, `sudo apt-get -o DPkg::Lock::Timeout=180 install -y dotnet-sdk-${versionUtils.getMajorMinor(mockVersion, acquisitionContext.eventStream, acquisitionContext)}`);
         }
     }).timeout(standardTimeoutTime);
 
@@ -185,7 +191,7 @@ Microsoft.NETCore.App 7.0.5 [/usr/lib/dotnet/shared/Microsoft.NETCore.App]`, std
         if (shouldRun)
         {
             await provider.uninstallDotnet(mockVersion, installType);
-            assert.equal(mockExecutor.attemptedCommand, 'sudo apt-get -o DPkg::Lock::Timeout=180 remove -y dotnet-sdk-7.0');
+            assert.equal(mockExecutor.attemptedCommand, `sudo apt-get -o DPkg::Lock::Timeout=180 remove -y dotnet-sdk-${versionUtils.getMajorMinor(mockVersion, acquisitionContext.eventStream, acquisitionContext)}`);
         }
     }).timeout(standardTimeoutTime);
 
@@ -194,7 +200,7 @@ Microsoft.NETCore.App 7.0.5 [/usr/lib/dotnet/shared/Microsoft.NETCore.App]`, std
         if (shouldRun)
         {
             await provider.upgradeDotnet(mockVersion, installType);
-            assert.equal(mockExecutor.attemptedCommand, 'sudo apt-get -o DPkg::Lock::Timeout=180 upgrade -y dotnet-sdk-7.0');
+            assert.equal(mockExecutor.attemptedCommand, `sudo apt-get -o DPkg::Lock::Timeout=180 upgrade -y dotnet-sdk-${versionUtils.getMajorMinor(mockVersion, acquisitionContext.eventStream, acquisitionContext)}`);
         }
     }).timeout(standardTimeoutTime * 1000);
 });