+++ vscode-dotnet-runtime-library/src/test/unit/WinMacGlobalInstaller.test.ts
@@ -21,6 +21,7 @@ const standardTimeoutTime = 100000;
 
 suite('Windows & Mac Global Installer Tests', function ()
 {
+    this.retries(0);
     const mockVersion = '7.0.306';
     const mockUrl = 'https://download.visualstudio.microsoft.com/download/pr/4c0aaf08-3fa1-4fa0-8435-73b85eee4b32/e8264b3530b03b74b04ecfcf1666fe93/dotnet-sdk-7.0.306-win-x64.exe';
     const mockHash = '';